run examples through test
